Basic documentation and summer cleaning.
