+++ Example/Source/Component/Examples/ProgressView/ProgressViewController.swift
@@ -12,4 +12,10 @@ final class ProgressViewController: ControllerBase<Void, ProgressViewRootView> {
     override init() {
         super.init(title: "Progress view")
     }
+
+    override func afterInit() {
+        navigationItem.rightBarButtonItem = UIBarButtonItem(barButtonSystemItem: .camera) {
+            ApplicationTheme.selector.select(theme: ApplicationTheme.selector.currentTheme == .day ? .night : .day)
+        }
+    }
 }
