+++ Sources/Tokenizer/Contexts/PropertyContext.swift
@@ -7,6 +7,9 @@
 
 import Foundation
 
+/**
+ * Context connected to a property (i.e. a field inside a UI element).
+ */
 public struct PropertyContext: DataContext, HasParentContext {
     public let parentContext: DataContext
     public let property: Property
