+++ Sources/Tokenizer/Elements/ComponentDefinition.swift
@@ -22,6 +22,9 @@ public enum AccessModifier: String {
     case `internal`
 }
 
+/**
+ * Contains the structure of a Component's file.
+ */
 public struct ComponentDefinition: XMLElementDeserializable, UIContainer, UIElementBase, StyleContainer, ComponentDefinitionContainer {
     public var type: String
     public var isRootView: Bool
@@ -56,6 +59,11 @@ public struct ComponentDefinition: XMLElementDeserializable, UIContainer, UIElem
     }
 
     #if canImport(UIKit)
+    /**
+     * **[LiveUI]** Adds a `UIView` to the passed self.
+     * - parameter subview: view to be added as a subview
+     * - parameter toInstanceOfSelf: parent to which the view should be added
+     */
     public func add(subview: UIView, toInstanceOfSelf: UIView) {
         toInstanceOfSelf.addSubview(subview)
     }
@@ -79,6 +87,11 @@ public struct ComponentDefinition: XMLElementDeserializable, UIContainer, UIElem
         properties = try PropertyHelper.deserializeSupportedProperties(properties: View.availableProperties, in: node)
     }
 
+    /**
+     * Try to deserialize a `ComponentDefinition` from an XML element.
+     * - parameter node: XML element to try to parse
+     * - returns: if not thrown, then the `ComponentDefinition` representing the XML element passed
+     */
     public static func deserialize(_ node: SWXMLHash.XMLElement) throws -> ComponentDefinition {
         return try ComponentDefinition(node: node, type: node.value(ofAttribute: "type"))
     }
