+++ Sources/Tokenizer/Layout/Constraints/Condition.swift
@@ -22,26 +22,6 @@ struct ConditionError: Error {
     }
 }
 
-public struct InterfaceState {
-    public let interfaceIdiom: InterfaceIdiom
-    public let horizontalSizeClass: InterfaceSizeClass
-    public let verticalSizeClass: InterfaceSizeClass
-    public let viewOrientation: ViewOrientation
-    public let rootDimensions: (width: Float, height: Float)
-
-    public init(interfaceIdiom: InterfaceIdiom,
-                horizontalSizeClass: InterfaceSizeClass,
-                verticalSizeClass: InterfaceSizeClass,
-                rootDimensions: (width: Float, height: Float)) {
-
-        self.interfaceIdiom = interfaceIdiom
-        self.horizontalSizeClass = horizontalSizeClass
-        self.verticalSizeClass = verticalSizeClass
-        self.rootDimensions = rootDimensions
-        self.viewOrientation = ViewOrientation(width: rootDimensions.width, height: rootDimensions.height)
-    }
-}
-
 public enum ConditionBinaryOperation {
     case and
     case or
@@ -75,6 +55,13 @@ public enum ConditionUnaryOperation {
     case negation
 }
 
+/**
+ * Description of a condition as a binary tree where:
+ * **statement** is the node itself
+ * **unary** is a node with one child
+ * **binary** is a node with two children
+ * - TODO: Make the `ConditionStatement` generic as to allow for arbitrary uses of this enum.
+ */
 public indirect enum Condition {
     case statement(ConditionStatement)
     case unary(ConditionUnaryOperation, Condition)
@@ -214,215 +201,6 @@ public indirect enum Condition {
     }
 }
 
-public enum ConditionStatement {
-    case interfaceIdiom(InterfaceIdiom)
-    case sizeClass(SizeClassType, InterfaceSizeClass)
-    case interfaceSizeClass(InterfaceSizeClass)
-    case orientation(ViewOrientation)
-    case trueStatement
-    case falseStatement
-    case number(Float)
-    case dimensionType(DimensionType)
-
-    public var isComparable: Bool {
-        switch self {
-        case .number, .dimensionType:
-            return true
-        case .interfaceIdiom, .sizeClass, .interfaceSizeClass, .orientation, .trueStatement, .falseStatement:
-            return false
-        }
-    }
-
-    init?(identifier: String) {
-        let lowerIdentifier = identifier.lowercased()
-        switch lowerIdentifier {
-        case "phone", "iphone":
-            self = .interfaceIdiom(.phone)
-        case "pad", "ipad":
-            self = .interfaceIdiom(.pad)
-        case "tv", "appletv":
-            self = .interfaceIdiom(.tv)
-        case "carplay":
-            self = .interfaceIdiom(.carPlay)
-        case "horizontal":
-            self = .sizeClass(.horizontal, .unspecified)
-        case "vertical":
-            self = .sizeClass(.vertical, .unspecified)
-        case "landscape":
-            self = .orientation(.landscape)
-        case "portrait":
-            self = .orientation(.portrait)
-        case "compact":
-            self = .interfaceSizeClass(.compact)
-        case "regular":
-            self = .interfaceSizeClass(.regular)
-        case "false":
-            self = .falseStatement
-        case "true":
-            self = .trueStatement
-        case "width":
-            self = .dimensionType(.width)
-        case "height":
-            self = .dimensionType(.height)
-        default:
-            return nil
-        }
-    }
-
-    func mergeWith(statement: ConditionStatement) -> ConditionStatement? {
-        if case .sizeClass(let sizeClass, _) = self,
-            case .interfaceSizeClass(let interfaceSizeClass) = statement {
-            return .sizeClass(sizeClass, interfaceSizeClass)
-        } else if case .interfaceSizeClass(let interfaceSizeClass) = self,
-            case .sizeClass(let sizeClass, _) = statement {
-            return .sizeClass(sizeClass, interfaceSizeClass)
-        } else {
-            return nil
-        }
-    }
-
-    #if canImport(UIKit)
-    func numberValue(from traits: UITraitHelper) -> Float {
-        switch self {
-        case .interfaceIdiom, .sizeClass, .orientation, .trueStatement, .falseStatement, .interfaceSizeClass:
-            fatalError("Requested `numberValue` from a logical condition statement.")
-        case .number(let value):
-            return value
-        case .dimensionType(let type):
-            switch type {
-            case .width:
-                return traits.viewRootSize(.width)
-            case .height:
-                return traits.viewRootSize(.height)
-            }
-        }
-    }
-
-    func evaluate(from traits: UITraitHelper) throws -> Bool {
-        switch self {
-        case .interfaceIdiom(let idiom):
-            return traits.device(idiom.runtimeValue)
-        case .sizeClass(let type, let sizeClass):
-            if type == .horizontal {
-                return traits.size(horizontal: sizeClass.runtimeValue)
-            } else {
-                return traits.size(vertical: sizeClass.runtimeValue)
-            }
-        case .orientation(let orientation):
-            return traits.orientation(orientation)
-        case .trueStatement:
-            return true
-        case .falseStatement:
-            return false
-        case .number:
-            throw ConditionError("Can't evaluate number only.")
-        case .interfaceSizeClass:
-            throw ConditionError("Can't evaluate interfaceSizeClass only.")
-        case .dimensionType:
-            throw ConditionError("Can't evaluate dimensionType only.")
-        }
-    }
-    #endif
-}
-
-public enum InterfaceIdiom {
-    case pad
-    case phone
-    case tv
-    case carPlay
-    case unspecified
-
-    var description: String {
-        switch self {
-        case .pad:
-            return "pad"
-        case .phone:
-            return "phone"
-        case .tv:
-            return "tv"
-        case .carPlay:
-            return "carPlay"
-        case .unspecified:
-            return "unspecified"
-        }
-    }
-
-    #if canImport(UIKit)
-    var runtimeValue: UIUserInterfaceIdiom {
-        switch self {
-        case .pad:
-            return .pad
-        case .phone:
-            return .phone
-        case .tv:
-            return .tv
-        case .carPlay:
-            return .carPlay
-        case .unspecified:
-            return .unspecified
-        }
-    }
-    #endif
-}
-
-public enum InterfaceSizeClass {
-    case compact
-    case regular
-    case unspecified
-
-    var description: String {
-        switch self {
-        case .compact:
-            return "compact"
-        case .regular:
-            return "regular"
-        case .unspecified:
-            return "unspecified"
-        }
-    }
-
-    #if canImport(UIKit)
-    var runtimeValue: UIUserInterfaceSizeClass {
-        switch self {
-        case .compact:
-            return .compact
-        case .regular:
-            return .regular
-        case .unspecified:
-            return .unspecified
-        }
-    }
-    #endif
-}
-
-public enum SizeClassType {
-    case horizontal
-    case vertical
-
-    var description: String {
-        switch self {
-        case .horizontal:
-            return "horizontal"
-        case .vertical:
-            return "vertical"
-        }
-    }
-}
-
-public enum DimensionType {
-    case width
-    case height
-
-    var description: String {
-        switch self {
-        case .width:
-            return "width"
-        case .height:
-            return "height"
-        }
-    }
-}
-
 // MARK: - Generator Extensions
 extension Condition {
     public func generateSwift(viewName: String) -> String {
@@ -457,51 +235,6 @@ extension Condition {
     }
 }
 
-extension ConditionStatement {
-    public func generateSwift(viewName: String) -> String {
-        switch self {
-        case .trueStatement:
-            return "true"
-        case .falseStatement:
-            return "false"
-        case .sizeClass(let sizeClassType, let viewInterfaceSize):
-            return "\(viewName).traits.size(\(sizeClassType.description): .\(viewInterfaceSize.description))"
-        case .interfaceIdiom(let interfaceIdiom):
-            return "\(viewName).traits.device(.\(interfaceIdiom.description))"
-        case .orientation(let orientation):
-            return "\(viewName).traits.orientation(.\(orientation.description))"
-        case .number(let number):
-            return String(number)
-        case .dimensionType(let dimensionType):
-            return "\(viewName).traits.viewRootSize(.\(dimensionType.description))"
-        case .interfaceSizeClass:
-            fatalError("Swift condition code generation reached an undefined point.")
-        }
-    }
-
-    public func generateXML() -> String {
-        switch self {
-        case .trueStatement:
-            return "true"
-        case .falseStatement:
-            return "false"
-        case .sizeClass(let sizeClassType, let viewInterfaceSize):
-            return "\(sizeClassType.description) == \(viewInterfaceSize.description)"
-        case .interfaceIdiom(let interfaceIdiom):
-            return interfaceIdiom.description
-        case .orientation(let orientation):
-            return orientation.description
-        case .number(let number):
-            return String(number)
-        case .dimensionType(let dimensionType):
-            return dimensionType.description
-        case .interfaceSizeClass:
-            fatalError("XML condition code generation reached an undefined point.")
-        }
-    }
-}
-
-// MARK: - Helper extensions for generating code
 extension ConditionBinaryOperation {
     var swiftRepresentation: String {
         switch self {
