+++ Sources/Tokenizer/Properties/Descriptions/ElementAssignablePropertyDescription.swift
@@ -7,6 +7,9 @@
 
 import Foundation
 
+/**
+ * Property description describing a property using a single XML element.
+ */
 public struct ElementAssignablePropertyDescription<T: ElementSupportedPropertyType>: TypedPropertyDescription {
     public typealias ValueType = T
 
@@ -15,11 +18,22 @@ public struct ElementAssignablePropertyDescription<T: ElementSupportedPropertyTy
     public let swiftName: String
     public let key: String
 
+    /**
+     * Get a property using the dictionary passed.
+     * - parameter properties: **[name: property]** dictionary to search in
+     * - returns: found property's value if found, nil otherwise
+     */
     public func get(from properties: [String: Property]) -> T? {
         let property = getProperty(from: properties)
         return property?.value
     }
 
+    /**
+     * Set a property's value from the dictionary passed.
+     * A new property is created if no property is found.
+     * - parameter value: value to be set to the property
+     * - parameter properties: **[name: property]** dictionary to search in
+     */
     public func set(value: T, to properties: inout [String: Property]) {
         var property: ElementAssignableProperty<T>
         if let storedProperty = getProperty(from: properties) {
@@ -31,10 +45,20 @@ public struct ElementAssignablePropertyDescription<T: ElementSupportedPropertyTy
         setProperty(property, to: &properties)
     }
 
+    /**
+     * Gets a property from the **[name: property]** dictionary passed or nil.
+     * - parameter dictionary: properties dictionary
+     * - returns: found property or nil
+     */
     private func getProperty(from dictionary: [String: Property]) -> ElementAssignableProperty<T>? {
         return dictionary[dictionaryKey()] as? ElementAssignableProperty<T>
     }
 
+    /**
+     * Inserts the property passed into the dictionary of properties.
+     * - parameter property: property to insert
+     * - parameter dictionary: **[name: property]** dictionary to insert into
+     */
     private func setProperty(_ property: Property, to dictionary: inout [String: Property]) {
         dictionary[dictionaryKey()] = property
     }
@@ -45,7 +69,6 @@ public struct ElementAssignablePropertyDescription<T: ElementSupportedPropertyTy
 }
 
 extension ElementAssignablePropertyDescription: ElementPropertyDescription {
-
     public func materialize(element: XMLElement) throws -> Property {
         let materializedValue = try T.materialize(from: element)
