+++ Sources/Tokenizer/Properties/Descriptions/ElementControlStatePropertyDescription.swift
@@ -11,6 +11,9 @@ import Foundation
 import UIKit
 #endif
 
+/**
+ * Property description describing a property using a single XML element with a control state.
+ */
 public struct ElementControlStatePropertyDescription<T: ElementSupportedPropertyType>: TypedPropertyDescription, ControlStatePropertyDescriptionMarker {
     public typealias ValueType = T
 
@@ -18,16 +21,34 @@ public struct ElementControlStatePropertyDescription<T: ElementSupportedProperty
     public let name: String
     public let key: String
 
+    /**
+     * Checks whether the passed attribute name can be handled by this `PropertyDescription`.
+     * - parameter attributeName: attribute name to be checked
+     * - returns: whether or not the passed attribute name can be handled
+     */
     public func matches(attributeName: String) -> Bool {
         let resolvedAttributeName = namespace.resolvedAttributeName(name: name)
         return attributeName == resolvedAttributeName || attributeName.hasPrefix("\(resolvedAttributeName).")
     }
 
+    /**
+     * Get a property using the dictionary passed for the passed control state.
+     * - parameter properties: **[name: property]** dictionary to search in
+     * - parameter state: `[ControlState]` to use when getting a property in the dictionary
+     * - returns: found property's value if found for the passed control state, nil otherwise
+     */
     public func get(from properties: [String: Property], for state: [ControlState]) -> T? {
         let property = getProperty(from: properties, for: state)
         return property?.value
     }
 
+    /**
+     * Set a property's value from the dictionary passed for the passed control state.
+     * A new property is created if no property corresponding to the control state is found.
+     * - parameter value: value to be set to the property
+     * - parameter properties: **[name: property]** dictionary to search in
+     * - parameter state: `[ControlState]` to find within the dictionary
+     */
     public func set(value: T, to properties: inout [String: Property], for state: [ControlState]) {
         var property: ElementControlStateProperty<T>
         if let storedProperty = getProperty(from: properties, for: state) {
@@ -39,10 +60,22 @@ public struct ElementControlStatePropertyDescription<T: ElementSupportedProperty
         setProperty(property, to: &properties, for: state)
     }
 
+    /**
+     * Gets a property from the **[name: property]** dictionary passed for the passed state or nil.
+     * - parameter dictionary: properties dictionary
+     * - parameter state: `[ControlState]` for which to search
+     * - returns: found property or nil
+     */
     private func getProperty(from dictionary: [String: Property], for state: [ControlState]) -> ElementControlStateProperty<T>? {
         return dictionary[dictionaryKey(for: state)] as? ElementControlStateProperty<T>
     }
 
+    /**
+     * Inserts the property passed into the dictionary of properties for passed state.
+     * - parameter property: property to insert
+     * - parameter dictionary: **[name: property]** dictionary to insert into
+     * - parameter state: `[ControlState]` to insert with
+     */
     private func setProperty(_ property: Property, to dictionary: inout [String: Property], for state: [ControlState]) {
         dictionary[dictionaryKey(for: state)] = property
     }
