+++ Sources/Tokenizer/Properties/Descriptions/MultipleAttributeAssignablePropertyDescription.swift
@@ -11,6 +11,9 @@ import Foundation
 import UIKit
 #endif
 
+/**
+ * Property description describing a property using multiple XML attributes.
+ */
 public struct MultipleAttributeAssignablePropertyDescription<T: MultipleAttributeSupportedPropertyType>: TypedPropertyDescription {
     public typealias ValueType = T
 
@@ -19,11 +22,22 @@ public struct MultipleAttributeAssignablePropertyDescription<T: MultipleAttribut
     public let swiftName: String
     public let key: String
 
+    /**
+     * Get a property using the dictionary passed.
+     * - parameter properties: **[name: property]** dictionary to search in
+     * - returns: found property's value if found, nil otherwise
+     */
     public func get(from properties: [String: Property]) -> T? {
         let property = getProperty(from: properties)
         return property?.value
     }
 
+    /**
+     * Set a property's value from the dictionary passed.
+     * A new property is created if no property is found.
+     * - parameter value: value to be set to the property
+     * - parameter properties: **[name: property]** dictionary to search in
+     */
     public func set(value: T, to properties: inout [String: Property]) {
         var property: MultipleAttributeAssignableProperty<T>
         if let storedProperty = getProperty(from: properties) {
@@ -35,10 +49,20 @@ public struct MultipleAttributeAssignablePropertyDescription<T: MultipleAttribut
         setProperty(property, to: &properties)
     }
 
+    /**
+     * Gets a property from the **[name: property]** dictionary passed or nil.
+     * - parameter dictionary: properties dictionary
+     * - returns: found property or nil
+     */
     private func getProperty(from dictionary: [String: Property]) -> MultipleAttributeAssignableProperty<T>? {
         return dictionary[dictionaryKey()] as? MultipleAttributeAssignableProperty<T>
     }
 
+    /**
+     * Inserts the property passed into the dictionary of properties.
+     * - parameter property: property to insert
+     * - parameter dictionary: **[name: property]** dictionary to insert into
+     */
     private func setProperty(_ property: Property, to dictionary: inout [String: Property]) {
         dictionary[dictionaryKey()] = property
     }
@@ -49,7 +73,6 @@ public struct MultipleAttributeAssignablePropertyDescription<T: MultipleAttribut
 }
 
 extension MultipleAttributeAssignablePropertyDescription: MultipleAttributePropertyDescription where T: MultipleAttributeSupportedPropertyType {
-
     public func materialize(attributes: [String : String]) throws -> Property {
         let namePrefixLength = namespace.resolvedAttributeName(name: name).count + 1
