+++ Sources/Tokenizer/Properties/MultipleAttributeAssignableProperty.swift
@@ -11,6 +11,10 @@ import Foundation
 import UIKit
 #endif
 
+/**
+ * Typed property that is obtained from multiple XML attributes.
+ * Its main usage is for `struct`s that need multiple `init` parameters.
+ */
 public struct MultipleAttributeAssignableProperty<T: MultipleAttributeSupportedPropertyType>: TypedProperty {
     public var namespace: [PropertyContainer.Namespace]
     public var name: String
@@ -21,6 +25,11 @@ public struct MultipleAttributeAssignableProperty<T: MultipleAttributeSupportedP
         return namespace.resolvedAttributeName(name: name)
     }
 
+    /**
+     * - parameter target: UI element to be targetted with the property
+     * - parameter context: property context to use
+     * - returns: Swift `String` representation of the property application on the target
+     */
     public func application(on target: String, context: PropertyContext) -> String {
         let namespacedTarget = namespace.resolvedSwiftName(target: target)
         return "\(namespacedTarget).\(description.swiftName) = \(value.generate(context: context.child(for: value)))"
@@ -33,6 +42,11 @@ public struct MultipleAttributeAssignableProperty<T: MultipleAttributeSupportedP
     #endif
 
     #if canImport(UIKit)
+    /**
+     * Try to apply the property on an object using the passed property context.
+     * - parameter object: UI element to apply the property to
+     * - parameter context: property context to use
+     */
     public func apply(on object: AnyObject, context: PropertyContext) throws {
         let key = description.key
         let selector = Selector("set\(key.capitalizingFirstLetter()):")
