+++ Sources/Tokenizer/Properties/Types/Implementation/Shadow.swift
