+++ Sources/Tokenizer/Themes/ApplicationDescription.swift
@@ -7,12 +7,22 @@
 
 import Foundation
 
+/**
+ * Structure containing themed resources.
+ * An example would be themed XML element **<Fonts>** (and its innards) inside an **<Application>** XML element.
+ */
 public struct ThemeContainer<T: AttributeSupportedPropertyType>: XMLElementDeserializable {
     public typealias ItemName = String
 
     private var defaultItems: [ItemName: T] = [:]
     private var themedItems: [ApplicationDescription.ThemeName: [ItemName: T]] = [:]
 
+    /**
+     * Given a theme and an item name returns its property type.
+     * - parameter theme: theme to use
+     * - parameter item: item to get using the passed theme
+     * - returns: property type corresponding to passed theme and item if present, otherwise nil
+     */
     public subscript(theme theme: String, item item: String) -> T? {
         guard let themeItems = themedItems[theme] else {
             return defaultItems[item]
@@ -21,6 +31,9 @@ public struct ThemeContainer<T: AttributeSupportedPropertyType>: XMLElementDeser
         return themeItems[item] ?? defaultItems[item]
     }
 
+    /**
+     * A `Set` of all themed items inside the theme container.
+     */
     public var allItemNames: Set<ItemName> {
         let defaultItemNames = defaultItems.keys
         let themedItemNames = themedItems.values.flatMap { $0.keys }
@@ -42,11 +55,20 @@ public struct ThemeContainer<T: AttributeSupportedPropertyType>: XMLElementDeser
         }
     }
 
+    /**
+     * Tries to deserialize the `ThemeContainer` from an XML element.
+     * - parameter element: XML element to parse
+     * - returns: if not thrown, the `ThemeContainer` structure
+     */
     public static func deserialize(_ element: XMLElement) throws -> ThemeContainer<T> {
         return try ThemeContainer(node: element)
     }
 }
 
+/**
+ * Structure describing an **<Application>** XML element containing **<Themes>**, **<Colors>**, **<Images>**, and **<Fonts>**
+ * that are to be themeable inside the application.
+ */
 public struct ApplicationDescription: XMLElementDeserializable {
     public static let defaultThemeName = "none"
 
@@ -88,6 +110,9 @@ public struct ApplicationDescription: XMLElementDeserializable {
         try! validate()
     }
 
+    /**
+     * Tries to validate the application description, throws if it's invalid.
+     */
     public func validate() throws {
         // MARK:- Validation
         guard themes.contains(defaultTheme) else {
@@ -95,10 +120,20 @@ public struct ApplicationDescription: XMLElementDeserializable {
         }
     }
 
+    /**
+     * Tries to deserialize the `ApplicationDescription` from an XML element.
+     * - parameter element: XML element to parse
+     * - returns: if not thrown, the `ApplicationDescription` structure
+     */
     public static func deserialize(_ element: XMLElement) throws -> ApplicationDescription {
         return try ApplicationDescription(node: element)
     }
 
+    /**
+     * Checks if the passed value is themed.
+     * - parameter value: the value to check
+     * - returns: `Optional` value, the original value without the prefix if check succeeded, nil otherwise
+     */
     public static func themedValueName(value: String) -> String? {
         let themePrefix = "theme."
         if value.hasPrefix(themePrefix) {
