+++ Sources/Tokenizer/Util/Float+CGFloat.swift
@@ -5,6 +5,7 @@
 //  Created by Matouš Hýbl on 23/04/2017.
 //  Copyright © 2017 Brightify. All rights reserved.
 //
+
 import CoreGraphics
 import Foundation
