+++ Sources/Tokenizer/Util/Swift+Compatibility.swift
@@ -13,5 +13,4 @@ public extension Sequence {
         return try self.flatMap(transform)
     }
 }
-
 #endif
